requirements.txt
langchain-core>=0.3.37
langchain-community
langchain
langgraph>=0.0.10
google-generativeai
langchain-google-genai
requests
python-dotenv
bs4
langgraph-checkpoint-sqlite
